Found hopefully correcto computation for additional effort

+++ src/time_integration/methods_PERKMulti.jl
@@ -192,7 +192,8 @@ mutable struct PERK_Multi_Integrator{RealT<:Real, uType, Params, Sol, F, Alg, PE
   max_lvl::Int64
   du_ode_hyp::uType # TODO: Not best solution since this is not needed for hyperbolic problems
   dtRef::RealT
-  AddRHSCalls::Int64
+  #AddRHSCalls::Int64
+  AddRHSCalls::Float64
 end
 
 # Forward integrator.stats.naccept to integrator.iter (see GitHub PR#771)
@@ -684,7 +685,8 @@ function solve(ode::ODEProblem, alg::PERK_Multi;
                 level_info_boundaries_acc, level_info_boundaries_orientation_acc,
                 level_info_mortars_acc, 
                 level_u_indices_elements,
-                t0, -1, n_levels, min_level, max_level, du_ode_hyp, dt, 0)
+                # CARE: Highest level = 9 
+                t0, -1, n_levels, min_level, max_level, du_ode_hyp, dt, 0.0)
             
   # initialize callbacks
   if callback isa CallbackSet
